pyproject.toml
[project]
name = "prompt-tuning"
version = "0.1.0"
description = "Prompt tuning with PEFT"
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "accelerate>=0.30.0",
    "datasets>=2.19.0",
    "peft>=0.11.0",
    "python-dotenv>=1.0.1",
    "transformers>=4.41.0",
    "torch>=2.4.0",
    "torchvision",
    "torchaudio",
    "numpy<2",
]

# -------------------------
# PyTorch CUDA index
# -------------------------
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }